degrib: fix int overflow. Fixes https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=4315. Credit to OSS Fuzz
git-svn-id: https://svn.osgeo.org/gdal/trunk@40747 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/grib/degrib/degrib/clock.c
@@ -2464,7 +2464,7 @@ double Clock_AddMonthYear (double refTime, int incrMonth, int incrYear)
    Clock_Epoch2YearDay (totDay, &day, &year);
    month = Clock_MonthNum (day, year);
    day = day - Clock_NumDay (month, 1, year, 1) + 1;
-   d_remain = refTime - totDay * 3600 * 24.0;
+   d_remain = refTime - (double)totDay * 3600 * 24.0;
 
    /* Add the month */
    if (incrMonth != 0) {
